Update patch set 2
Patch Set 2:

Hi Kunal,

What we really want is for the RAID module to release and close the base bdevs when the RAID volume is destroyed during shutdown.  This will allow the bdev layer to make intelligent decisions about the order to destroy bdevs - it can start with bdevs that are not claimed.  But if the RAID module doesn't release the claim on the base bdevs, the bdev layer cannot use this strategy.

I do have an alternate suggestion for this - I am planning to add a new module callback, where modules (like RAID) can be notified when app shutdown is starting.  Would this simplify things?

Thanks,

-Jim

Patch-set: 2
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>