Update patch set 60
Patch Set 60:

Have not yet addressed comments from Ben or Lance on 59, this update was about fixing some data corruption issues under load w/large IOs.  From dist list email:

FYI I just updated the patch and fixed all but one data corruption issue that I could find. Also addressed the concern with large arrays as best as I could and commented on why they are what they are.  Wrt why I had to move them back to stack variables as opposed to keeping them in the channel, I’ll see if I can explain here as this was one of the sources of data corruption.

I need to assure, of course, that I finish all crypto operations on a read/write before completing the bdev io.  It’s a bit different for each, I’ll briefly explain and highlight where the corruption issue was by using channel structure based arrays. Note that the arrays in question are arrays of pointers to crypto operation structures and the address of the array is submitted to the crypto driver for both burst encryption and decryption, this is an important point:

READ: (decrypt in place)
•	Submit the read (via spdk_bdev_readv_blocks()) to the underlying bdev supplying my own callback.
•	In the read callback, assuming success, call the general _crypto_operation() function using same src and dst (the buffers described by the original read)
•	The function will submit all crypto operations before existing even if it has to manually call the poller to complete some of them in the event that the crypto device is full. Note that the device may have other operations outstanding from a previous IO when it is drained during submission. 
•	When the poller has completed all operations associated with a bdev_io, whether called from the submission routine or during the period run as a poller, the original read is completed.
•	When the poller is called in one of the two manners just described, it provides the address of an array of pointers to hold de-queued operations. If that array is located in the channel and the completion of this read was triggered while the poller still had more operations to de-queue, the subsequent call back into the poller, before things unwind, will supply the crypto driver the address of the channel array of pointers that may not be completely processed yet by the earlier poller that hasn’t completed yet. By putting the array on the stack they are independent.  This nesting can only ever be 2 deep by the way because it only happens on reads as just described.

WRITE: (encrypt in side buffer so we don’t mess with the original data given by the app)
•	Immediately call the _crypto_operation() function to encrypt the data in the src buffer. However, we allocate a new dst buffer (see comments in code on how this has to be done for now) to hold the encrypted data.  The new buffer is contiguous regardless of the IOV makeup of the original.
•	 The function will submit all crypto operations before existing even if it has to manually call the poller to complete some of them in the event that the crypto device is full. Note that the device may have other operations outstanding from a previous IO when it is drained during submission. 
•	When the poller has completed all operations associated with a bdev_io, whether called from the submission routine or during the period run as a poller, it will issue a new write to the underlying bdev using spdk_bdev_writev_blocks() targeted at the underlying bdev using the contiguous encrypted buffer we allocated earlier.

I could probably put the submission array pointers in the channel, I don’t see any issues with that, but for consistency just put them on the stack also.

Now, the one data corruption I have yet to figure out. I can run bdevio and bdevperf (and UT of course) against:
- straight NVME
- straight malloc
- split malloc 2 disks, split 1 disk

With verify on QD from 2-512 and IO sizes 512 to 64K with no issues.  But when I use split + NVMe (one split disk or 2) I get data buffer mismatch on 64K IOs at QD of around 16 or higher. This is still a bit of a mystery to me….

Patch-set: 60
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>