Update patch set 2
Patch Set 2:

Curious, did you run into a situation where you needed this, debug of a specific issue or something? Or is this just something that you noticed and felt like should be added and if so under what scenarios would this be that useful? Just asking to learn :)

Patch-set: 2
CC: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>