Update patch set 60
Patch Set 60:

(copied from SPDK mailing list)

Any of the async operations cannot point to addresses that are in the channel (and certainly not the stack of course).  Are we sure that’s not happening?  If you can push your latest patch to GerritHub I can take a look.  The comment here was the part that concerned me:
 
When the poller is called in one of the two manners just described, it provides the address of an array of pointers to hold de-queued operations. If that array is located in the channel and the completion of this read was triggered while the poller still had more operations to de-queue, the subsequent call back into the poller, before things unwind, will supply the crypto driver the address of the channel array of pointers that may not be completely processed yet by the earlier poller that hasn’t completed yet. By putting the array on the stack they are independent.  This nesting can only ever be 2 deep by the way because it only happens on reads as just described.

Patch-set: 60
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>