Update patch set 60
Patch Set 60:

(from mailing list)
Hi Jim-

It’s up there are of this morning.  Correct, there are no async operations that rely on earlier channel or stack variables – well AFAIK of course  Here’s maybe an easier to read sequence of events saying the same thing as that paragraph is trying to:

1) my submit() function is called
2) I call spdk_bdev_readv_blocks(), nothing special wrt crypto at this point
3) async callback for the read happens
4) from within the callback I call _crypto_operation
5) crypto operation provides the address to a stack array of pointers when it asks the crypto driver to enqueue
6) assume, for simplicity, that all ops an enqueued
7) assume some other IOs happen and pile up crypto operations in the crypto driver 
8) my crypto poller runs and provides the crypto driver the address of a stack var array of pointer to tell us what has been dequeuer
9) in the poller, we loop through the stack array
10) assume we get 80% through it and find enough to complete the bdev_io from step (1)
11) we sync call our internal “the crypto operation is done, do the next thing”
12) since this was a read we do a sync call to decrypt the data
13) the sync call enqueue a bunch of operations to the crypto driver
14) if the driver is full, which can happen pretty easily with large IOs, the enqueue path will now call the poller manually, remember we haven’t unwound from the sync call made forom the poller back in step 10
15) Now the poller needs to ask the crypto driver to fill a new array with completed operations. If it gave it anything other than a stack variable, the 20% of the operations that hadn’t completed in step 10 would get overwritten btu because it’s a stack variable they don’t.  This poller completes, unwinds
16) we end up back at step 10 where the first instance of the poller completes the other 20% and marches on

I looked for a simpler way to handle this but the complexity is really driven by the fact that I chose to drain the crypto driver in the enqueue path if it was full.  There are other options but I’ll let you digest and identify and issues with this one first 

Thanks for reviewing Jim, appreciate it. This is more complex than I had anticipated once we went to LBA as IV.

Patch-set: 60
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>