Update patch set 1
Patch Set 1:

Hi Darek,

Good question.  The end goal here is that during shutdown, the bdev layer picks the next bdev to unregister by finding bdevs that are not claimed.  So we can't wait for module_fini to run - we need the lvol module to release the claim on the underlying bdev (that holds the lvolstore) once the last lvol associated with it is unregistered.

This shouldn't apply to virtio-scsi, since even if there are no LUNs, you don't need to claim a bdev to create the virtio-scsi connection.

Your point is taken though - I'll update the commit message with some more of these details on the end goal.

-Jim

Patch-set: 1
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>