Update patch set 3
Patch Set 3: Code-Review+2

I do think it's clearer if it's named "update" instead of "get" - I rechecked the code to remind myself why we'd call "get" but not use the return value.  Maybe a good reader would do that anyways.

I'm fine with this patch as-is for now, and changing the name in a later patch, only so you don't have to repush this whole series.

Patch-set: 3
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2