Update patch set 8
Patch Set 8:

Hi Shuhei,
I agree with you. Whether we remove raid bdev first or base bdev, both are handled as part of shutdown in the raid module code. But I got from Jim that there is a requirement where in shutdown case only if we remove raid_bdev, it should release the all base bdev descriptors and the claim. This was not possible earlier because bdev does not tell to module that shutdown is in progress. So I think Jim added shutdown notification for this reason.

Patch-set: 8
Reviewer: Kunal Sablok <1014253@d5d70762-12d0-45a1-890d-524b12d3f735>