Update patch set 3
Patch Set 3:

It's great initiative to add the tables to easy to read, from the table, we can see the test result categorized by test file name, and is it possible to add the column for the test file test result. And in the long run, it's really helpful to add the test cases, like which case pass, which case failed, and this will also need to add tag to the test case in the log file i think.

Patch-set: 3
CC: qun wan <1011885@d5d70762-12d0-45a1-890d-524b12d3f735>