Update patch set 9
Patch Set 9:

(4 comments)

Historically we've tried to maintain backward compatibility for the configuration file and the RPCs, as much as we can. We mark things as deprecated and give users warnings for at least one release before dropping old formats. Figuring out how to maintain backward compatibility in the config file and RPC calls for this change is going to be a major challenge though. I need to think through the best way to approach this and I don't have an answer just yet, but I wanted to post my other comments in the mean time.

Patch-set: 9
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>