Update patch set 4
Patch Set 4:

Hi Shuhei,

If blobstore is on a RAID-0, and the RAID-0 is on multiple NVMe namespaces, then the sequence of operations would be like this for a 8MB write:

1) Blobstore send only first 4MB write to RAID-0 (sequential)
2) RAID-0 will break up the 4MB write into separate commands per member disk (batch)
3) Once #1 and #2 are complete, blobstore send next 4MB write to RAID-0 (sequential)
4) RAID-0 will break up the 4MB write into separate commands per member disk (batch)

I think this is best - this sequential operation only happens when the I/O size is very big - and even then there is still some amount of batching to the member disks.

What do you think?

-Jim

Patch-set: 4
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>