Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Could we make this the responsibility of the bdev driver itself?  I mean, the passthrough driver should know if its already claimed the bdev.

Sure, each and every driver could be managing its own claims, doing very much the same refcount-based thing. At the same time, it's simpler to do once in the bdev layer. 

Furthermore, it's also a nice example of claims being not that straightforward replacement for vbdev->base_bdevs since, as it turns out in this example, the same vbdev might be claiming the same bdev multiple times.

Summing up, that's something I needed to make multiple multipath vbdevs work with the same base bdev. This could be easily dropped/moved, but the key observation here is that we have first to decide which mechanism - claims or base_bdevs/vbdevs is used as a foundation of the bdev tree management - and then focus on that one.

Patch-set: 1
Reviewer: Andrey Kuzmin <1014177@d5d70762-12d0-45a1-890d-524b12d3f735>