blobstore: Make IOV split efficient by keeping current position of IOV
Current IOV split operation in _spdk_rw_iov_split_next(ctx, bserrno)
is not so efficient because index and offset of IOV are moved to the
current position for every call.

Holding current position of index and offset of the original IOV
is more efficient and is done in this patch.

This algorithm of IOV split operation in blobstore will be utilized
in RAID bdev in upcoming patches.

Change-Id: I61ee998e841feed1a5adace2c066f0fa89ed4da5
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>