bdev/raid: Refactor split IO operation to extend it to IOV
Current strip index is usable to count outstanding and pending IOs.
But if we refer blobstore's IOV split operation, using start and
size may be natural and easier to support IOV split operation.

This change also simplifies retry operation.

Operation in the helper function get_curr_base_bdev_index() became
simple and was used only in a place. Hence remove it in this patch
too.

Change-Id: I20966810d927aaf05d3dfad7e4c993c2186bbc4e
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>