Update patch set 4
Patch Set 4:

I pushed a patch https://review.gerrithub.io/#/c/spdk/spdk/+/422942/. But this is for the most simple case (single IOV) and can not be used to this patch yet.

I've worked on IOV pool management, but this is not easy for me and I may add this item in the upcoming community meeting.

Patch-set: 4