Update patch set 9
Patch Set 9: Code-Review+2

I'd like to keep this as a uint64_t for now.  We've set the limit to 16 trace groups, so a uint16_t would work, but we may want to increase the number of trace groups in the future.  Or if a user wanted to test with 32 trace groups for some reason, they could just change one #define and not need to change anything else.

Patch-set: 9
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2