event: Remove reactor sleep functionality
Reactors had the ability to go to sleep if no work was being
executed, but any poller counted as work. While the thread
library is refactored, just remove this functionality.

This patch leaves in the configuration options for this
feature, even though they now do nothing. They can be
hooked up again when the functionality is added again.

Change-Id: I497c5d466dc5034debf8c9762fa62950ac35a7ce
Signed-off-by: Ben Walker <benjamin.walker@intel.com>