Update patch set 5
Patch Set 5: Code-Review-1

Hi Avinash - I have a few architectural comments about how the code is laid out that will help inform some of the implementation here.

First, the NVMe-oF target is divided into two distinct parts: the library and the application. The library is intended to be a generic piece of code that can be embedded into a range of applications. The primary impact this has is that the library attempts to not enforce any specific threading model, but rather places restrictions on what can be done from each thread simultaneously. The code for the library is all in lib/nvmf and the API is in include/spdk/nvmf.h. The rest of the code in the NVMe-oF target (mostly in lib/event/subsystems/nvmf) is all part of what we call the "application". Technically, this is just an example of how to use the library, but we test it and fix bugs as if it were production quality. This application has to choose a threading model (and chooses one thread per core, with one poll group per thread).

Given the above, the scheduling of qpairs onto poll groups is an application-level decision. That means that all of the changes you're making here (minus one, which I'll outline below) should be entirely contained in lib/event/subsystems/nvmf/*.

The only exception is that the application needs some way to ask the library which host IP is associated with a given qpair. You'll need to add a function to the library to return this information (i.e. add a function to include/spdk/nvmf.h). SPDK defines an abstraction of a network address that is fabric agnostic called spdk_nvme_transport_id. Your function should allow the user to pass one of those in and fill it out appropriately. For an RDMA-based qpair, the traddr contains the IP address and the trsvcid contains the port.

I hope that helps. Thanks for your effort here!

Patch-set: 5
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1