event: make timer reinsertion more precise
Features like bdev QoS rely on being invoked at a
specific frequency.  So modify the reactor to reinsert
the timer based on when the timeout was scheduled to run -
not the actual time when the timeout ran.  This accounts
for any delays associated with events, pollers or other
timers that may have slightly delayed a timeout from
occuring exactly when requested.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: If5bd0f5f4cfe692604f2ef5592ff9d2ef03e6340