[RFC] env_dpdk/pci: map all vfio resources before device init
We've relied on the fact that DPDK maps all its memory
to vfio unconditionally so we don't need to register
our additional regions before a device is fully initialized.
That's why we could successfully call spdk_vtophys() inside
a PCI device-specific init.

We're about to take over the responsibility of registering
DPDK's memory, so make sure we do that before that device
-specific init.

Change-Id: If1d3d310073ffa5f9ad285ea05433fe29cbaef4a
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>