lib/nvmf: add the nvmf qpair to the group with valid channels
In the case that the subsystem in the related poll group has
NULL IO channel assigned due to some problem like out of resource,
for example, the NVMe SSD hardware itself has limited number
of IO qpairs. The subsystems in the particular poll group
could have zero valid channels. In this case, when a new
NVMe-oF qpair assigned to this poll group, it will first check
whether the poll group has valid channels. If not, find the
next poll group to add the NVMe-oF qpair which is in the
round robin mode.

Change-Id: Iedee2a6375e48eb7bf899cfb0542c565c7ebd231
Signed-off-by: GangCao <gang.cao@intel.com>