lib/nvmf: complete the request immediately when subsystem is inactive
In the case that subsystem is in the inactive state and never has
the chance to go to active state, the request shall not be queued,
otherwise this request will be timedout eventually at the initiator
side. Here it just completes this request immediately.

Proper error is returned that the initiator could behave accordingly
to this error code.

Change-Id: Iedee2a6375e48eb7bf899cfb0542c565c7ebd231
Signed-off-by: GangCao <gang.cao@intel.com>