Update patch set 4
Patch Set 4:

Detailed information can be found here https://github.com/spdk/spdk/issues/395

The basic scenario is for the case where 32 cores are assigned to the target and there is NULL IO QPair on particular core due to the limitation of the NVMe SSD (i.e., P3700).

The error case could be general that the related Subsystem is in the inactive state and could never go to the Active state to drain the Queued I/Os. The original behavior has uncertainty and based on how Initiator handles. For example, just hang some time and timed out to reset the target controller to drop the outstanding I/Os. The fix is to return immediately with the proper error here.

Patch-set: 4
Reviewer: GangCao <1011207@d5d70762-12d0-45a1-890d-524b12d3f735>